subsystem: fs: fcb: add option to calculate crc when writing content
This commit also adds the f_crc_precalculate option to the fcb struct.
By enabling crc precalulation, the flash entry content is not read
back to calculate the CRC in fcb_append_finish. This is beneficial for
applications that needs increased write throughput.

When the crc precalculation is enabled fcb_append_crc must be called
after every successful write of entry content to append the content
to the fcb_entry CRC. This is written to flash when calling
fcb_append_finish.

TODO: Test and doc update.

Signed-off-by: Eivind Jølsgard <eivind.jolsgard@nordicsemi.no>