[nrf fromtree] drivers: serial: uart_nrfx_uarte: coexisting async and interrupt API
This commit fixes an issue with the nrfx uarte driver to allow the
async and interrupt driven UART APIs to coexist on different uart
instances. As both APIs cannot be used simultaneously for a given
instance, there is no need to handle
CONFIG_UART_EXCLUSIVE_API_CALLBACKS in this driver.

(cherry picked from commit 66069b3)

Signed-off-by: Eivind Jølsgard <eivind.jolsgard@nordicsemi.no>